split resource manager
